Move sorted data file sorting helpers
