+++ engine/src/main/java/org/hestiastore/index/sorteddatafile/sort/MergedEntryIterator.java
@@ -1,4 +1,4 @@
-package org.hestiastore.index.sorteddatafile;
+package org.hestiastore.index.sorteddatafile.sort;
 
 import java.util.ArrayList;
 import java.util.Comparator;
